Refactor - Add LabMemberController - edit routes
