+++ app/Http/Middleware/MustBeAdmin.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Middleware;
 
+use App\Models\User;
 use Closure;
 
 class MustBeAdmin
@@ -16,7 +17,7 @@ class MustBeAdmin
     public function handle($request, Closure $next)
     {
         $user = $request->user();
-        if ($user && $user->type == 'ADMIN') {
+        if ($user && $user->type == User::ADMIN) {
             return $next($request);
         } 
         abort(403, 'No way.');
