+++ app/Http/Requests/PracticeRequest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class PracticeRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'name'       => 'required|unique:practices',
+            'admin_name' => 'required',
+            'email'      => 'required|email|unique:users',
+            'password'   => 'required|confirmed'
+        ];
+    }
+}
