+++ notes.md
@@ -1,10 +1,28 @@
 # Test project FarmLab
 
-1. LabResult
+1. User
+    + LabMemberController
+        * CRUD on farm lab team members
+    + VetController
+        * CRUD on vets
+
 2. Practice
-3. File
-4. User
+    + PracticeController
+        * CRUD on practices
+
+3. LabResult
+    + LabResultController
+        * CRUD on lab results
+
+4. File
+    + FileController
+        * CRUD on file
 
 
+# User
 
-
+- User can be of type:
+    1. farmlab admin
+    2. farmlab team member
+    3. practice admin
+    4. practice vet
