+++ resources/views/practice/create.blade.php
@@ -5,21 +5,30 @@
         <div class="row justify-content-center">
             <div class="col-md-8">
                 <div class="card">
-                    <div class="card-header">Add new team member</div>
+                    <div class="card-header">Add new practice</div>
 
                     <div class="card-body">
-                        <form method="POST" action="{{ route('farmlab.store') }}">
+                        <form method="POST" action="{{ route('practice.store') }}">
                             @csrf
 
                             <div class="form-group row">
-                                <label for="name" class="col-sm-4 col-form-label text-md-right">Username</label>
+                                <label for="name" class="col-sm-4 col-form-label text-md-right">Practice name</label>
                                 <div class="col-md-6">
                                     <input name="name" type="text" class="form-control" id="name" required>
                                 </div>
                             </div>
 
                             <div class="form-group row">
-                                <label for="email" class="col-sm-4 col-form-label text-md-right">Email</label>
+                                <label for="admin_name" class="col-sm-4 col-form-label text-md-right">Practice admin
+                                    name</label>
+                                <div class="col-md-6">
+                                    <input name="admin_name" type="text" class="form-control" id="admin_name" required>
+                                </div>
+                            </div>
+
+                            <div class="form-group row">
+                                <label for="email" class="col-sm-4 col-form-label text-md-right">Practice admin
+                                    email</label>
                                 <div class="col-md-6">
                                     <input name="email" type="email" class="form-control" id="email" required>
                                 </div>
@@ -44,16 +53,15 @@
                             <div class="form-group row mb-0">
                                 <div class="col-md-8 offset-md-4">
                                     <button type="submit" class="btn btn-primary">
-                                        Add member
+                                        Add practice
                                     </button>
                                 </div>
                             </div>
+
                         </form>
                     </div>
                 </div>
             </div>
             @include ('layouts.errors')
         </div>
-    </div>
-
 @endsection
