+++ resources/views/vets/index.blade.php
