+++ resources/views/vets/show.blade.php
